added evolution experience config with amount per 1 experience system

+++ Assets/_Game/Scripts/GamePlay/Player/Modules/Experience/ExperienceModule.cs
@@ -26,10 +26,10 @@ public class ExperienceModule: IDisposable, IResource
         _mouthModule.OnFoodPointsAchieved += AddExperience;
     }
 
-    private void AddExperience(int amount)
+    private void AddExperience(float amount)
     {
-        OnExperienceGained?.Invoke(amount);
-        UpdateExperience(amount);
+        OnExperienceGained?.Invoke((int)amount);
+        UpdateExperience((int)amount);
     }
 
     private void UpdateExperience(int amount)
